Deleted unused CommandHelper. Moved Command and Alias from otbproject.commands.loader to otbproject.commands package
