+++ src/main/java/com/github/otbproject/otbproject/commands/Aliases.java
@@ -1,7 +1,6 @@
 package com.github.otbproject.otbproject.commands;
 
 import com.github.otbproject.otbproject.App;
-import com.github.otbproject.otbproject.commands.loader.Alias;
 import com.github.otbproject.otbproject.database.DatabaseWrapper;
 import com.github.otbproject.otbproject.users.UserLevel;
