+++ src/main/java/com/github/otbproject/otbproject/proc/CommandProcessor.java
@@ -3,8 +3,8 @@ package com.github.otbproject.otbproject.proc;
 import com.github.otbproject.otbproject.App;
 import com.github.otbproject.otbproject.commands.Aliases;
 import com.github.otbproject.otbproject.commands.Commands;
-import com.github.otbproject.otbproject.commands.loader.Alias;
-import com.github.otbproject.otbproject.commands.loader.Command;
+import com.github.otbproject.otbproject.commands.Alias;
+import com.github.otbproject.otbproject.commands.Command;
 import com.github.otbproject.otbproject.commands.parser.CommandResponseParser;
 import com.github.otbproject.otbproject.database.DatabaseWrapper;
 import com.github.otbproject.otbproject.users.UserLevel;
