+++ src/main/java/com/github/otbproject/otbproject/util/preload/PreloadComparator.java
@@ -2,8 +2,8 @@ package com.github.otbproject.otbproject.util.preload;
 
 import com.github.otbproject.otbproject.commands.Aliases;
 import com.github.otbproject.otbproject.commands.Commands;
-import com.github.otbproject.otbproject.commands.loader.Alias;
-import com.github.otbproject.otbproject.commands.loader.Command;
+import com.github.otbproject.otbproject.commands.Alias;
+import com.github.otbproject.otbproject.commands.Command;
 import com.github.otbproject.otbproject.database.DatabaseWrapper;
 import com.github.otbproject.otbproject.filters.BasicFilter;
 import com.github.otbproject.otbproject.filters.FilterGroup;
